[INLONG-9286][Agent] Adjust the time offset calculation function (#9288)
