+++ inlong-agent/agent-plugins/src/main/java/org/apache/inlong/agent/plugin/task/filecollect/LogFileCollectTask.java
@@ -80,6 +80,8 @@ public class LogFileCollectTask extends Task {
     public static final long DAY_TIMEOUT_INTERVAL = 2 * 24 * 3600 * 1000;
     public static final int CORE_THREAD_SLEEP_TIME = 1000;
     public static final int CORE_THREAD_MAX_GAP_TIME_MS = 60 * 1000;
+    public static final int CORE_THREAD_PRINT_TIME = 10000;
+    private long lastPrintTime = 0;
     private boolean retry;
     private long startTime;
     private long endTime;
@@ -221,6 +223,9 @@ public class LogFileCollectTask extends Task {
 
     @Override
     public String getTaskId() {
+        if (taskProfile == null) {
+            return "";
+        }
         return taskProfile.getTaskId();
     }
 
@@ -234,6 +239,10 @@ public class LogFileCollectTask extends Task {
         Thread.currentThread().setName("directory-task-core-" + getTaskId());
         running = true;
         while (!isFinished()) {
+            if (AgentUtils.getCurrentTime() - lastPrintTime > CORE_THREAD_PRINT_TIME) {
+                LOGGER.info("log file task running! taskId {}", getTaskId());
+                lastPrintTime = AgentUtils.getCurrentTime();
+            }
             coreThreadUpdateTime = AgentUtils.getCurrentTime();
             AgentUtils.silenceSleepInMs(CORE_THREAD_SLEEP_TIME);
             if (!initOK) {
@@ -274,7 +283,7 @@ public class LogFileCollectTask extends Task {
     private void scanExistingFile() {
         originPatterns.forEach((originPattern) -> {
             List<BasicFileInfo> fileInfos = scanExistingFileByPattern(originPattern);
-            LOGGER.info("scan {} get file count {}", originPattern, fileInfos.size());
+            LOGGER.info("taskId {} scan {} get file count {}", getTaskId(), originPattern, fileInfos.size());
             fileInfos.forEach((fileInfo) -> {
                 addToEvenMap(fileInfo.fileName, fileInfo.dataTime);
             });
@@ -299,7 +308,7 @@ public class LogFileCollectTask extends Task {
             long currentTime = System.currentTimeMillis();
             // only scan two cycle, like two hours or two days
             long offset = NewDateUtils.calcOffset("-2" + taskProfile.getCycleUnit());
-            startScanTime = currentTime - offset;
+            startScanTime = currentTime + offset;
             endScanTime = currentTime;
         }
         return FileScanner.scanTaskBetweenTimes(taskProfile, originPattern, startScanTime, endScanTime, retry);
